Clearing stuff and renaming the main file
