--- src/Resource.h
@@ -1,10 +1,0 @@
-#ifndef RESOURCE_H
-#define RESOURCE_H
-
-class Resource {
- public:
-  Resource( const std::string &resPath );
-  int resId;
-  std::string *resPath;
-};
-#endif
